+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fixed error handling / automatic retries of downloads in entrypoint script.
+
 ## [v2.7.0](https://github.com/allenai/beaker-gantry/releases/tag/v2.7.0) - 2025-06-24
 
 ### Added