+++ CHANGELOG.md
@@ -10,10 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `-f/--follow` option to `gantry logs`.
+- Added `--python-version` option to `gantry run` for overriding the default Python version to use.
 
 ### Changed
 
 - Changed default beaker image to [`petew/gantry`](https://beaker.allen.ai/orgs/ai2/workspaces/gantry-testing/images).
+
+### Fixed
+
+- Made installing/upgrading the `pip` package manager more robust, and added a check to ensure the pip version being used is in the active virtual environment.
 
 ## [v2.7.1](https://github.com/allenai/beaker-gantry/releases/tag/v2.7.1) - 2025-06-25